# backend/api/docs/swagger.yaml
definitions:
...
  github_com_imlargo_go-api-template_internal_dto.UpdateEvaluationAttemptRequest:
    properties:
      completed_at:
        type: string
      is_completed:
        type: boolean
      score:
        type: number
    type: object
...
  github_com_imlargo_go-api-template_internal_enums.ContentType:
    enum:
    - content
    - evaluation
    type: string
    x-enum-varnames:
    - ContentTypeContent
    - ContentTypeEvaluation
...
  github_com_imlargo_go-api-template_internal_enums.QuestionType:
    enum:
    - single_choice
    - multiple_choice
    type: string
    x-enum-varnames:
    - QuestionTypeSingle
    - QuestionTypeMultiple
  github_com_imlargo_go-api-template_internal_enums.UserRole:
    enum:
    - student
    - instructor
    - admin
    type: string
    x-enum-varnames:
    - UserRoleStudent
    - UserRoleInstructor
    - UserRoleAdmin
  github_com_imlargo_go-api-template_internal_models.Answer:
    properties:
      created_at:
        type: string
      id:
        type: integer
      isCorrect:
        type: boolean
      order:
        type: integer
      question:
        allOf:
        - $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Question'
        description: Relaciones
      question_id:
        type: integer
      text:
        type: string
      updated_at:
        type: string
    type: object
  github_com_imlargo_go-api-template_internal_models.AttemptAnswer:
    properties:
      is_correct:
        type: boolean
      points:
        type: integer
      question_id:
        type: integer
      selected_answer_ids:
        items:
          type: integer
        type: array
    type: object
  github_com_imlargo_go-api-template_internal_models.Content:
    properties:
      body:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      mediaUrl:
        type: string
      module:
        allOf:
        - $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Module'
        description: Relaciones
      module_id:
        type: integer
      order:
        type: integer
      title:
        type: string
      type:
        $ref: '#/definitions/github_com_imlargo_go-api-template_internal_enums.ContentType'
      updated_at:
        type: string
      user_progress:
        items:
          $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.UserProgress'
        type: array
    type: object
  github_com_imlargo_go-api-template_internal_models.Course:
    properties:
      created_at:
        type: string
      description:
        type: string
      enrollments:
        items:
          $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Enrollment'
        type: array
      id:
        type: integer
      image_url:
        type: string
      module_count:
        type: integer
      modules:
        description: Relaciones
        items:
          $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Module'
        type: array
      short_description:
        type: string
      student_count:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  github_com_imlargo_go-api-template_internal_models.Enrollment:
    properties:
      completed_at:
        type: string
      course:
        $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Course'
      course_id:
        type: integer
      created_at:
        type: string
      enrolled_at:
        type: string
      id:
        type: integer
      progress:
        description: porcentaje 0-100
        type: number
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.User'
        description: Relaciones
      user_id:
        type: integer
    type: object
  github_com_imlargo_go-api-template_internal_models.Evaluation:
    properties:
      created_at:
        type: string
      description:
        type: string
      evaluation_attempts:
        items:
          $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.EvaluationAttempt'
        type: array
      id:
        type: integer
      max_attempts:
        type: integer
      module:
        allOf:
        - $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Module'
        description: Relaciones
      module_id:
        type: integer
      order:
        type: integer
      passing_score:
        type: integer
      question_count:
        type: integer
      questions:
        items:
          $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Question'
        type: array
      time_limit:
        description: en minutos
        type: integer
      title:
        type: string
      type:
        $ref: '#/definitions/github_com_imlargo_go-api-template_internal_enums.ContentType'
      updated_at:
        type: string
      user_progress:
        items:
          $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.UserProgress'
        type: array
    type: object
  github_com_imlargo_go-api-template_internal_models.EvaluationAttempt:
    properties:
      answers:
        items:
          $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.AttemptAnswer'
        type: array
      created_at:
        type: string
      evaluation:
        $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Evaluation'
      evaluation_id:
        type: integer
      id:
        type: integer
      passed:
        type: boolean
      score:
        type: integer
      started_at:
        type: string
      submitted_at:
        type: string
      time_spent:
        description: en minutos
        type: integer
      total_points:
        type: integer
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.User'
        description: Relaciones
      user_id:
        type: integer
    type: object
...
  github_com_imlargo_go-api-template_internal_models.Module:
    properties:
      contents:
        items:
          $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Content'
        type: array
      course:
        allOf:
        - $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Course'
        description: Relaciones
      course_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      evaluations:
        items:
          $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Evaluation'
        type: array
      id:
        type: integer
      order:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_imlargo_go-api-template_internal_models.Question:
    properties:
      answers:
        items:
          $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Answer'
        type: array
      created_at:
        type: string
      evaluation:
        allOf:
        - $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Evaluation'
        description: Relaciones
      evaluation_id:
        type: integer
      explanation:
        type: string
      id:
        type: integer
      points:
        type: integer
      text:
        type: string
      type:
        $ref: '#/definitions/github_com_imlargo_go-api-template_internal_enums.QuestionType'
      updated_at:
        type: string
    type: object
  github_com_imlargo_go-api-template_internal_models.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      enrollments:
        description: Relaciones
        items:
          $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Enrollment'
        type: array
      fullname:
        type: string
      id:
        type: integer
      role:
        $ref: '#/definitions/github_com_imlargo_go-api-template_internal_enums.UserRole'
      updated_at:
        type: string
    type: object
  github_com_imlargo_go-api-template_internal_models.UserProgress:
    properties:
      attempts:
        type: integer
      completed_at:
        type: string
      content:
        $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Content'
      content_id:
        description: puede ser Content o Evaluation
        type: integer
      course:
        $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Course'
      course_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      module:
        $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.Module'
      module_id:
        type: integer
      score:
        type: integer
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.User'
        description: Relaciones
      user_id:
        type: integer
    type: object
  github_com_imlargo_go-api-template_internal_responses.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact:
    name: Default
    url: https://default.dev
  description: Backend service template
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Go api
  version: "1.0"
paths:
...
  /api/v1/evaluation-attempts/{id}:
    get:
      description: Get an evaluation attempt by its ID
      parameters:
      - description: Attempt ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.EvaluationAttempt'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get evaluation attempt
      tags:
      - evaluation-attempts
    patch:
      consumes:
      - application/json
      description: Update an evaluation attempt by ID
      parameters:
      - description: Attempt ID
        in: path
        name: id
        required: true
        type: integer
      - description: Attempt data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_imlargo_go-api-template_internal_dto.UpdateEvaluationAttemptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Attempt updated successfully
          schema:
            $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.EvaluationAttempt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_imlargo_go-api-template_internal_responses.ErrorResponse'
        "404":
          description: Attempt not found
          schema:
            $ref: '#/definitions/github_com_imlargo_go-api-template_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_imlargo_go-api-template_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update evaluation attempt
      tags:
      - evaluation-attempts
  /api/v1/evaluation-attempts/{id}/score:
    post:
      description: Score an evaluation attempt (admin only)
      parameters:
      - description: Attempt ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.EvaluationAttempt'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Score evaluation attempt
      tags:
      - evaluation-attempts
  /api/v1/evaluation-attempts/{id}/submit:
    post:
      consumes:
      - application/json
      description: Submit answers for an evaluation attempt
      parameters:
      - description: Attempt ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.EvaluationAttempt'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Submit evaluation attempt
      tags:
      - evaluation-attempts
  /api/v1/evaluation-attempts/start:
    post:
      consumes:
      - application/json
      description: Start a new evaluation attempt for a user
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.EvaluationAttempt'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Start evaluation attempt
      tags:
      - evaluation-attempts
...
  /api/v1/users/{userId}/evaluations/{evaluationId}/attempts:
    get:
      description: Get all attempts for a user and evaluation
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Evaluation ID
        in: path
        name: evaluationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_imlargo_go-api-template_internal_models.EvaluationAttempt'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get user evaluation attempts
      tags:
      - evaluation-attempts
  /api/v1/users/{userId}/evaluations/{evaluationId}/can-attempt:
    get:
      description: Check if a user can attempt an evaluation
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Evaluation ID
        in: path
        name: evaluationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Check if user can attempt evaluation
      tags:
      - evaluation-attempts
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  PushApiKey:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"